Offsets for centered version
$.anystretch("full.jpg",{
    offsetX:-216,
    offsetY:-108
});
See demo in examples/demo-fullscreen.html
Offsets are always relative to the center of the full sized image but
aware of the current scaling. So a point at (100,50) from the center
gets projected to (200,100) from the center if the image is scaled
twice. The scale algorithm is simplified and more general now. As I
didn`t care for the left/right aligned versions I have not included
them for now.